Created BaseMousePointer. Renaming of prefabs and classes
