+++ Assets/MixedRealityToolkit.SDK/Features/UX/Meshes/Cursors/Materials/SpatialMousePointer.mat
