+++ Assets/MixedRealityToolkit.SDK/Features/UX/Prefabs/Pointers/ScreenSpaceMousePointer.prefab.meta
