+++ Assets/MixedRealityToolkit.SDK/Features/UX/Prefabs/Pointers/SpatialMousePointer.prefab.meta
