+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/Pointers/BaseMousePointer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f03e21f9376f35e41b60506838cbfeb9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 93d49a6ae68a4f6ca0fea653caaa74fc, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
