+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/Pointers/ScreenSpaceMousePointer.cs.meta
